split source code from test code.
